Create Lab 1 Part 1
Write your first hello world java program! Your goal is to create a program which can be run over the command line, and prints “Hello World!” to the screen. If you complete this with time to spare, consider making it more complex. Instead of saying “World”, ask the user for her name and print “Hello {{NAME}}”, where {{NAME}} is a placeholder for the name the user provides!

Example Run-Time
> Please enter your name:
> Elliott
> Hello, Elliott!